Add layers 3a / 3b / 4 test infra + scenario stubs (refs #16)
- Split crates/maestro-daemon into lib.rs + thin main.rs to expose
  pub async fn serve(addr, repo_root) -> Result<DaemonHandle> for tests.
- Layer 3a: TestDaemon helper (tests/common/mod.rs) + smoke_daemon.rs
  spawn a real daemon on an ephemeral port over a tempdir, no mocking.
- Layer 3b: frontend/playwright.config.ts boots the daemon via webServer;
  e2e/smoke.spec.ts asserts header + "Daemon: connected" status.
- Layer 4: tests/smoke.sh boots the daemon, asserts /runs and /pipelines
  return JSON, / returns HTML, and the JS asset 200s.
- Layer 5: docs/testing/scenarios/{run-minimal,edit-and-save}.md skeletons
  (content lands with #18 and #17 respectively).
- CI: rust job runs smoke.sh; new e2e job installs Chromium + runs Playwright.
